Manually apply MAC address when supplied (fixes #1)
If specified (via `--mac-address`), Docker only applies a user-selected
MAC address when moving the interface into the container's network
namespace. This means that the initial run of `udhcpc` will use the
random veth MAC address, breaking any statically configured leases
(since this is the IP address that Docker will assign to the container).